progress: break apart existing handcalcs into new structure
